+++ Source/ChessDev/QuestSystem/QuestStep.h
@@ -18,7 +18,7 @@ public:
     FString StepDescription;
 
     UPROPERTY(EditAnywhere, BlueprintReadWrite)
-    float StepProgress;
+    int StepProgress;
 
     UPROPERTY(EditAnywhere, BlueprintReadWrite)
     UTrigger* StepTrigger;
